DIS-37: Decompose AgentManager into telemetry, persona-reviews, feedback modules
Extracts ~1700 LOC out of `apps/server/src/agents/manager.ts` (4967 → 3368)
into focused, unit-testable modules. AgentManager keeps its public surface
as a thin façade so all existing callers continue to work without changes:

- `agents/errors.ts` — `AgentError` (broken out so the new modules can throw
  the canonical error type without circling back through manager.ts).
- `agents/types.ts` — shared agent record + status/phase types.
- `agents/telemetry.ts` — `getActivitySummary`, `getAgentHistory`,
  `getFeedbackSummary`, `listMedia` plus their result types.
- `agents/persona-reviews.ts` — review CRUD, `completePersonaReview` state
  machine, `submitReviewResolution`, `cancelReviewRecheck`, recheck context,
  and `resolveProgressPingStatus`.
- `agents/feedback.ts` — feedback submit/list/update/count queries and the
  `respondsToFeedbackId` linkage check.

manager.ts re-exports the moved types and `AgentError` so existing imports
(`mcp-handlers.ts`, `http-helpers.ts`, `ui-events.ts`, etc.) keep resolving.

Pure refactor — no behaviour changes. All vitest + Playwright suites pass
unchanged.
