+++ apps/server/src/agents/errors.ts
@@ -1,0 +1,8 @@
+export class AgentError extends Error {
+  statusCode: number;
+
+  constructor(message: string, statusCode: number) {
+    super(message);
+    this.statusCode = statusCode;
+  }
+}
